customer-entity-library/src/main/java/com/jrmcdonald/ext/spring/security/SpringSecurityAuthenticationFacade.java
package com.jrmcdonald.ext.spring.security;

import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;

public class SpringSecurityAuthenticationFacade implements AuthenticationFacade {

    @Override
    public Authentication getAuthentication() {
        return SecurityContextHolder.getContext().getAuthentication();
    }

    @Override
    public String getCustomerId() {
        return getAuthentication().getName();
    }
}
